[DEM-1204] Bluefors monitor server
